@@ -6,7 +6,7 @@ part of owlcpp project.
 #ifndef CATALOG_HPP_
 #define CATALOG_HPP_
 #include <limits>
-#include "boost/filesystem/path.hpp"
+#include <string>
 
 #include "owlcpp/io/config.hpp"
 #include "owlcpp/io/exception.hpp"
@@ -103,8 +103,6 @@ private:
  and add them to the catalog
  @param cat catalog;
  @param path symbolic path pointing to local file or directory;
- any type implicitly convertible to boost::filesystem::path can be used,
- e.g., std::string, const char*
  @param recurse if true add to catalog files located in sub-directories
  @param search_depth once ontologyIRI declaration is found, stop searching for
  versionIRI declaration after @b search_depth triples
@@ -115,7 +113,7 @@ private:
 *******************************************************************************/
 OWLCPP_IO_DECL std::size_t add(
          Catalog& cat,
-         boost::filesystem::path const& path,
+         const std::string& path,
          const bool recurse = false,
          const std::size_t search_depth = std::numeric_limits<std::size_t>::max()
 );
