@@ -6,12 +6,13 @@ part of owlcpp project.
 #ifndef OWLCPP_IO_SOURCE
 #define OWLCPP_IO_SOURCE
 #endif
+#include "fsystem.hpp"
+
 #include "owlcpp/io/catalog.hpp"
-
-#include "boost/filesystem.hpp"
-
 #include "owlcpp/io/read_ontology_iri.hpp"
 #include "owlcpp/rdf/print_node.hpp"
+
+#include <boost/throw_exception.hpp>
 
 namespace owlcpp {
 /*
@@ -31,15 +32,15 @@ namespace{
 /*
 *******************************************************************************/
 inline std::size_t add_to_catalog(
-         boost::filesystem::path const& path,
+         const std::string& path,
          Catalog& cat,
          const std::size_t search_depth,
          const bool ignore_errors
 ) {
-   const boost::filesystem::path cp = canonical(path);
+   const fs::path cp = fs::canonical(path);
    std::pair<std::string,std::string> pair;
    try{
-      pair = read_ontology_iri(cp, search_depth);
+      pair = read_ontology_iri(cp.string(), search_depth);
    } catch(Input_err const&) {
       //ignore
       return 0;
@@ -49,7 +50,7 @@ inline std::size_t add_to_catalog(
       BOOST_THROW_EXCEPTION(
                Input_err()
                << Input_err::msg_t("ontologyIRI not found")
-               << Input_err::str1_t(path.string())
+               << Input_err::str1_t(path)
       );
    }
    return cat.insert_doc(pair.first, cp.string(), pair.second).second ? 1 : 0;
@@ -64,7 +65,7 @@ std::size_t add_to_catalog(
    std::size_t n = 0;
    for( ; i1 != i2; ++i1 ) {
       if( is_regular_file(*i1) )
-         n += add_to_catalog(i1->path(), cat, search_depth, true);
+         n += add_to_catalog(i1->path().string(), cat, search_depth, true);
    }
    return n;
 }
@@ -75,24 +76,26 @@ std::size_t add_to_catalog(
 *******************************************************************************/
 std::size_t add(
          Catalog& cat,
-         boost::filesystem::path const& path,
+         const std::string& path,
          const bool recurse,
          const std::size_t search_depth
          ) {
-   if( ! exists(path) ) BOOST_THROW_EXCEPTION(
+    if (!fs::exists(path)) {
+        BOOST_THROW_EXCEPTION(
             Input_err()
             << Input_err::msg_t("not found")
-            << Input_err::str1_t(path.string())
-   );
-   if( is_directory(path) ) {
+            << Input_err::str1_t(path)
+        );
+    }
+   if( fs::is_directory(path) ) {
       if( recurse ) {
-         boost::filesystem::recursive_directory_iterator i1(path), i2;
+         fs::recursive_directory_iterator i1(path), i2;
          return add_to_catalog(i1, i2, cat, search_depth);
       } else {
-         boost::filesystem::directory_iterator i1(path), i2;
+         fs::directory_iterator i1(path), i2;
          return add_to_catalog(i1, i2, cat, search_depth);
       }
-   } else if( is_regular_file(path) ) {
+   } else if( fs::is_regular_file(path) ) {
       return add_to_catalog(path, cat, search_depth, false);
    }
    return 0;
